fix(self-scan): split test_skill_package.py -- crossed the god-file rule
CI's authoritative self-scan (real failure, reproduced in a clean
short-path clone, not the worktree doctor/repowise false-negative
seen elsewhere this session) failed diagnosis.hospital with
"architecture gate failure":

  Quality: 4423 -> 4326 (quality_degraded)
  God files: 33 -> 34; new since baseline: tests/test_skill_package.py
  (loc 425, functions 34; rule: functions>25&&loc>400)

The attestation test additions from this PR (8 methods across two
review rounds) pushed the file over both thresholds at once. Split
the 8 self-contained gh_cli_version/verify_build_provenance tests into
tests/test_skill_attestation.py -- a natural seam, no shared state
with the rest of the file beyond the load_bootstrap_module() helper
(duplicated, not imported, matching the existing test file's style).

test_skill_package.py: 485 -> 401 raw lines, 26 test methods. Reran
the self-scan against the same clean clone: diagnosis.hospital now
succeeds/pass, no quality_degraded or god_files_increased violations,
no mention of the file. Also updated every place that invoked
test_skill_package.py by exact filename so the new file actually
runs in CI: ci.yml (both self-scan jobs), release.yml, skill-check.yml
(plus skill-check.yml's path-trigger filter). Reran code-intel repin
--write after editing the workflows -- clean, no stale pins.
