+++ .github/workflows/release.yml
@@ -54,6 +54,7 @@ jobs:
         run: |
           python tests/test_repository_layout.py -v
           python tests/test_skill_package.py -v
+          python tests/test_skill_attestation.py -v
 
       - name: Setup Rust
         shell: pwsh
