+++ tests/test_skill_package.py
@@ -290,90 +290,6 @@ class SkillPackageTests(unittest.TestCase):
             expected = hashlib.sha256(archive.read_bytes()).hexdigest()
             self.assertEqual(bootstrap.sha256_file(archive), expected)
 
-    def test_gh_cli_version_none_when_gh_missing(self) -> None:
-        bootstrap = load_bootstrap_module()
-        with mock.patch.object(bootstrap.shutil, "which", return_value=None):
-            self.assertIsNone(bootstrap.gh_cli_version())
-
-    def test_gh_cli_version_parses_stdout(self) -> None:
-        bootstrap = load_bootstrap_module()
-        completed = mock.Mock(returncode=0, stdout="gh version 2.49.0 (2024-05-01)\n")
-        with mock.patch.object(
-            bootstrap.shutil, "which", return_value="/usr/bin/gh"
-        ), mock.patch.object(bootstrap.subprocess, "run", return_value=completed):
-            self.assertEqual(bootstrap.gh_cli_version(), (2, 49))
-
-    def test_gh_cli_version_none_on_timeout(self) -> None:
-        bootstrap = load_bootstrap_module()
-        with mock.patch.object(
-            bootstrap.shutil, "which", return_value="/usr/bin/gh"
-        ), mock.patch.object(
-            bootstrap.subprocess,
-            "run",
-            side_effect=bootstrap.subprocess.TimeoutExpired(cmd="gh", timeout=30),
-        ):
-            self.assertIsNone(bootstrap.gh_cli_version())
-
-    def test_verify_build_provenance_raises_on_timeout(self) -> None:
-        bootstrap = load_bootstrap_module()
-        with tempfile.TemporaryDirectory() as temp:
-            archive = Path(temp) / "release.zip"
-            archive.write_bytes(b"unused")
-            with mock.patch.object(
-                bootstrap, "gh_cli_version", return_value=(2, 49)
-            ), mock.patch.object(
-                bootstrap.subprocess,
-                "run",
-                side_effect=bootstrap.subprocess.TimeoutExpired(cmd="gh", timeout=30),
-            ):
-                with self.assertRaises(bootstrap.BootstrapError):
-                    bootstrap.verify_build_provenance(archive)
-
-    def test_verify_build_provenance_degrades_when_gh_missing(self) -> None:
-        bootstrap = load_bootstrap_module()
-        with tempfile.TemporaryDirectory() as temp:
-            archive = Path(temp) / "release.zip"
-            archive.write_bytes(b"unused")
-            with mock.patch.object(bootstrap, "gh_cli_version", return_value=None):
-                self.assertEqual(
-                    bootstrap.verify_build_provenance(archive), "degraded_missing_gh"
-                )
-
-    def test_verify_build_provenance_degrades_when_gh_too_old(self) -> None:
-        bootstrap = load_bootstrap_module()
-        with tempfile.TemporaryDirectory() as temp:
-            archive = Path(temp) / "release.zip"
-            archive.write_bytes(b"unused")
-            with mock.patch.object(bootstrap, "gh_cli_version", return_value=(2, 40)):
-                self.assertEqual(
-                    bootstrap.verify_build_provenance(archive), "degraded_missing_gh"
-                )
-
-    def test_verify_build_provenance_accepts_passing_attestation(self) -> None:
-        bootstrap = load_bootstrap_module()
-        with tempfile.TemporaryDirectory() as temp:
-            archive = Path(temp) / "release.zip"
-            archive.write_bytes(b"unused")
-            completed = mock.Mock(returncode=0, stdout="", stderr="")
-            with mock.patch.object(
-                bootstrap, "gh_cli_version", return_value=(2, 49)
-            ), mock.patch.object(bootstrap.subprocess, "run", return_value=completed):
-                self.assertEqual(bootstrap.verify_build_provenance(archive), "verified")
-
-    def test_verify_build_provenance_rejects_failing_attestation(self) -> None:
-        bootstrap = load_bootstrap_module()
-        with tempfile.TemporaryDirectory() as temp:
-            archive = Path(temp) / "release.zip"
-            archive.write_bytes(b"unused")
-            completed = mock.Mock(
-                returncode=1, stdout="", stderr="no matching attestations"
-            )
-            with mock.patch.object(
-                bootstrap, "gh_cli_version", return_value=(2, 49)
-            ), mock.patch.object(bootstrap.subprocess, "run", return_value=completed):
-                with self.assertRaises(bootstrap.BootstrapError):
-                    bootstrap.verify_build_provenance(archive)
-
     def test_download_rejects_archive_larger_than_limit(self) -> None:
         bootstrap = load_bootstrap_module()
